Add the ability to sample (high throughput) metrics
