--- test/.jshintrc
@@ -1,10 +1,0 @@
-{
-  // If you are using SublimeLinter, after modifying this config file, be sure to close and reopen
-  // the file(s) you were editing to see the changes take effect.
-
-  // Prohibit the use of undeclared variables.
-  "extends": "../.jshintrc",
-
-  // Make JSHint aware of Mocha globals.
-  "mocha": true
-}
